feat: wire guided sampling into caliberprop, share it via ladder.SamplePoints
caliberprop's tracker only called the raw BuildSamplePoints on each
direction's live reserve, so it never picked up the depletion-anchored
DgeoBps resampling liquidcore already had -- every cycle re-probed the
naive geo grid regardless of where the pool's real depletion knee sat.

liquidcore's samplePoints/estimateNearCapacityAmount only ever touched
generic entity.Pool.Extra/Reserves and ladder.Extra/Point, nothing
liquidcore-specific, so moves cleanly into the ladder package itself as
SamplePoints -- both liquidcore and caliberprop now call the same function
instead of each carrying their own copy. The guided-sampling test moves
with it (TestSamplePoints_GuidedByPreviousLadder), since it was already
only exercising the shared logic through a liquidcore *PoolTracker receiver
that added nothing of its own.
